+++ WebSite/Core/Handler/Api/SourceMaterialHandler.cs
@@ -145,7 +145,7 @@ namespace Backstage.Core.Handler
                 {
                     Avatar = user.Avatar,
                     UserName = user.UserName,
-                    Sex = user.Sex,
+                    Sex = (int)user.Sex,
                     Dateline = cm.CreateTime,
                     Message = cm.Content
                 };
